+++ demos/compose/src/main/java/androidx/media3/demo/compose/buttons/MinimalControls.kt
@@ -27,6 +27,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.media3.common.Player
+import androidx.media3.ui.compose.material3.buttons.NextButton
+import androidx.media3.ui.compose.material3.buttons.PreviousButton
 
 /**
  * Minimal playback controls for a [Player].