+++ RELEASENOTES.md
@@ -161,6 +161,9 @@
     *   Add `ContentFrame` Composable to `media3-ui-compose` which combines
         `PlayerSurface` management with aspect ratio resizing and covering with
         a shutter.
+    *   Create a new `media3-ui-compose-material3` module that will house
+        Material3-themed Composables.
+    *   Add NextButton and PreviousButton to `media3-ui-compose-material3`.
 *   Downloads:
 *   OkHttp extension:
 *   Cronet extension: